ctf-manager-bot/requirements.txt
aiostream==0.4.5
Babel==2.12.1
discord.py==2.3.2
maya==0.6.1
requests==2.31.0
